+++ CppDarkOrbitBot/BotUtils.h
@@ -18,6 +18,6 @@ Mat screenshotWindow(HWND hwnd);
 double calculateIoU(const cv::Rect& a, const cv::Rect& b);
 void applyNMS(const vector<Rect>& boxes, const vector<double>& scores, double nmsThreshold, vector<int>& indices);
 long long getCurrentMillis();
-int computeMillisPassed(long long start, long long finish);
+long long computeMillisPassed(long long start, long long finish);
 
 #endif
